# docs/swagger.yaml
basePath: /
definitions:
...
  profiles.CreateProfileRequest:
    properties:
      contact_person:
        maxLength: 100
        type: string
      date_birth:
        type: string
      fullname:
        maxLength: 100
        minLength: 2
        type: string
      is_mentor:
        type: boolean
      profile_description:
        maxLength: 1000
        type: string
      profile_picture:
        type: string
    required:
    - date_birth
    - fullname
    type: object
  rooms.CreateRoomRequest:
    properties:
      room_name:
        type: string
      schedule_id:
        type: integer
    required:
    - room_name
    - schedule_id
    type: object
  schedules.CreateScheduleRequest:
    properties:
      appointment_id:
        type: integer
      end_date:
        type: string
      start_date:
        type: string
    type: object
  schedules.DeleteScheduleRequest:
    properties:
      id:
        type: integer
    type: object
  schedules.UpdateScheduleRequest:
    properties:
      appointment_id:
        type: integer
      end_date:
        type: string
      id:
        type: integer
      start_date:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: Tim Developer
  description: Ini adalah dokumentasi API untuk aplikasi online meeting.
  termsOfService: http://swagger.io/terms/
  title: Online Meeting API
  version: "1.0"
paths:
...
  /profile:
    get:
      description: Mengambil data profile user yang login
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Ambil Profile
      tags:
      - Profile
    post:
      consumes:
      - application/json
      description: Membuat profil baru
      parameters:
      - description: Profile Data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/profiles.CreateProfileRequest'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Create Profile
      tags:
      - Profile
  /room/create:
    post:
      consumes:
      - application/json
      description: Membuat room meeting antara mentee dan mentor
      parameters:
      - description: Room Data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/rooms.CreateRoomRequest'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Buat Room Meeting
      tags:
      - Room
  /schedule/create:
    post:
      consumes:
      - application/json
      description: Menambahkan jadwal ketersediaan mentor
      parameters:
      - description: Schedule Data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schedules.CreateScheduleRequest'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Buat Jadwal
      tags:
      - Schedule
  /schedule/delete:
    delete:
      consumes:
      - application/json
      description: Menghapus jadwal mentor
      parameters:
      - description: Delete Payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schedules.DeleteScheduleRequest'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Hapus Jadwal
      tags:
      - Schedule
  /schedule/list:
    get:
      description: Menampilkan semua jadwal yang dibuat
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: List Jadwal
      tags:
      - Schedule
  /schedule/update:
    put:
      consumes:
      - application/json
      description: Mengubah jadwal ketersediaan mentor
      parameters:
      - description: Update Data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schedules.UpdateScheduleRequest'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Ubah Jadwal
      tags:
      - Schedule
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"